# src/swaggerOutput.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Projeto Trainee Backend - API",
    "description": "Projeto para capacitação trainee de backend, com o objetivo de criar uma API para gerenciar projetos e tarefas."
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearer_token": {
      "type": "http",
      "scheme": "bearer",
      "bearerFormat": "JWT",
      "name": "bearer_token",
      "in": "header"
    }
  },
  "consumes": [
    "application/json",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/task/{projectId}": {
      "get": {
        "tags": [
          "Task"
        ],
        "description": "Endpoint para retornar todas as tarefas de um projeto.",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do projeto."
          }
        ],
        "responses": {
          "200": {
            "description": "Tarefa criada com sucesso."
          },
          "400": {
            "description": "Erro ao criar a tarefa."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      },
      "post": {
        "tags": [
          "Task"
        ],
        "description": "Endpoint para criar uma tarefa",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do projeto que a tarefa pertence."
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "category": {
                  "example": "any"
                },
                "deadline": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tarefa criada com sucesso."
          },
          "400": {
            "description": "Erro ao criar tarefa."
          },
          "404": {
            "description": "Erro ao encontrar projeto."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/task/{taskId}": {
      "get": {
        "tags": [
          "Task"
        ],
        "description": "Endpoint para retornar uma tarefa.",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID da tarefa."
          }
        ],
        "responses": {
          "200": {
            "description": "Tarefas encontradas com sucesso."
          },
          "400": {
            "description": "Erro ao obter tarefa."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      },
      "put": {
        "tags": [
          "Task"
        ],
        "description": "Endpoint para atualizar uma tarefa.",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID da tarefa."
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "category": {
                  "example": "any"
                },
                "deadline": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tarefa atualizada com sucesso."
          },
          "400": {
            "description": "Erro ao atualizar tarefa."
          },
          "404": {
            "description": "Erro ao encontrar tarefa."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Task"
        ],
        "description": "Endpoint para excluir uma tarefa.",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID da terfa."
          }
        ],
        "responses": {
          "200": {
            "description": "Tarefa removida com sucesso."
          },
          "400": {
            "description": "Erro ao remover tarefa."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/task/featured-image/{taskId}": {
      "post": {
        "tags": [
          "Task"
        ],
        "description": "Endpoint para cadastrar uma imagem de destaque da tarefa",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID da tarefa."
          }
        ],
        "responses": {
          "200": {
            "description": "Imagem enviada com sucesso."
          },
          "400": {
            "description": "Erro ao enviar imagem."
          },
          "500": {
            "description": "Erro ao enviar imagem."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Task"
        ],
        "description": "Endpoint para excluir a imagem de destaque da tarefa",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID da tarefa."
          }
        ],
        "responses": {
          "200": {
            "description": "Imagem removida com sucesso."
          },
          "400": {
            "description": "Erro ao remover imagem."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/task/files/{taskId}": {
      "post": {
        "tags": [
          "Task"
        ],
        "description": "Endpoint para cadastrar arquivos da tarefa.",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID da tarefa."
          }
        ],
        "responses": {
          "200": {
            "description": "Arquivo enviado com sucesso."
          },
          "404": {
            "description": "Erro ao encontrar tarefa."
          },
          "500": {
            "description": "Erro ao enviar arquivo."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/task/complete/{taskId}": {
      "put": {
        "tags": [
          "Task"
        ],
        "description": "Endpoint para concluir uma tarefa.",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID da tarefa."
          }
        ],
        "responses": {
          "200": {
            "description": "Tarefa concluída com sucesso."
          },
          "400": {
            "description": "Erro ao concluir tarefa."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/task/{taskId}/{fileName}": {
      "delete": {
        "tags": [
          "Task"
        ],
        "description": "Endpoint para excluir o arquivo de uma tarefa.",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID da tarefa."
          },
          {
            "name": "fileName",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Nome do arquivo."
          }
        ],
        "responses": {
          "200": {
            "description": "Tarefa removida com sucesso."
          },
          "400": {
            "description": "Erro ao remover tarefa."
          },
          "404": {
            "description": "Erro ao encontrar projeto."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
...
  },
...
}